refactor(bookmark): rename lobster scripts with lobster_ prefix
All scripts that run as lobster pipeline steps are renamed with the
lobster_ prefix to make their execution context self-evident from the
filename. Non-lobster scripts (heartbeat, crons, utilities) unchanged.

Renamed (11 files):
  summarize.py                  → lobster_summarize.py
  filter_curation.py            → lobster_filter_curation.py
  generate_specs.py             → lobster_generate_specs.py
  prepare_topic_approval.py     → lobster_prepare_topic_approval.py
  ensure_topic_slot_available.py→ lobster_ensure_topic_slot_available.py
  finalize_review_cycle.py      → lobster_finalize_review_cycle.py
  compact_approval_preview.py   → lobster_compact_approval_preview.py
  create_tasks_from_proposals.py→ lobster_create_tasks_from_proposals.py
  resolve_topic_approval.py     → lobster_resolve_topic_approval.py
  list_review_candidates.py     → lobster_list_review_candidates.py
  ensure_non_empty.py           → lobster_ensure_non_empty.py

Updated references:
  - lobster yaml command: fields
  - rebuild_revised_approval.py path constants (also dropped dead BUILD_TASK_PROPOSALS ref)
  - lobster_list_review_candidates.py import
  - common.py adjacentComponents string
  - spec doc Key Scripts table (filenames + removed stale note)
