+++ agents/workflows/bookmark/lobster_compact_approval_preview.py
