+++ agents/workflows/bookmark/lobster_generate_specs.py
